Revert "Refactor settings view into modular tabs"
