--- components/app_views/settings/index.ts
@@ -1,7 +1,0 @@
-export { AppearanceTab } from "./appearance-tab";
-export { LayoutTab } from "./layout-tab";
-export { ComponentsTab } from "./components-tab";
-export { TypographyTab } from "./typography-tab";
-export { BehaviorTab } from "./behavior-tab";
-export { PreviewPanel } from "./preview-panel";
-
